+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.7.4] - 2026-04-08
+
+### Added
+- WA023 lint rule: "Rule entry is not a dict" (ERROR)
+- WA024 lint rule: "Phase value is not a list" (ERROR)
+
 ## [0.7.3] - 2026-04-07
 
 ### Added